+++ src/raw.c
@@ -213,6 +213,13 @@ union jls_version_u jls_raw_version(struct jls_raw_s * self) {
     return self->version;
 }
 
+struct jls_bkf_s * jls_raw_backend(struct jls_raw_s * self) {
+    if (self->backend.fd == -1) {
+        return NULL;
+    }
+    return &self->backend;
+}
+
 int32_t jls_raw_wr(struct jls_raw_s * self, struct jls_chunk_header_s * hdr, const uint8_t * payload) {
     JLS_LOGD3("wr @ %" PRId64 " : %d %s", jls_raw_chunk_tell(self), (int) hdr->tag, jls_tag_to_name(hdr->tag));
     RLE(jls_raw_wr_header(self, hdr));
@@ -456,6 +463,10 @@ int64_t jls_raw_chunk_tell_end(struct jls_raw_s * self) {
     }
     jls_raw_chunk_seek(self, starting_pos);
     return end_pos;
+}
+
+int64_t jls_raw_truncate(struct jls_raw_s * self) {
+    return jls_bk_truncate(&self->backend);
 }
 
 const char * jls_tag_to_name(uint8_t tag) {